up and runing

+++ wordpress/wordpress-entrypoint.sh
@@ -1,35 +1,33 @@
 #!/bin/bash
 
-MYSQL_DATABASE='wp_dp'
-MYSQL_USER='abdelatif'
-MYSQL_PASSWORD='secure123'
-MYSQL_HOST='mariadb'
-
-# until mysqladmin ping -h "$MYSQL_HOST" --user=$MYSQL_USER --password="$MYSQL_PASSWORD" --silent
-# do
-#     echo "Waiting for MySQL..."
-#     sleep 2
-# done
-# Ping the container until it responds (exit code 0)
 while true; do
     if ping -c 1 "$MYSQL_HOST" &> /dev/null; then
-        echo "Container is reachable!"
+        echo "MariaDB is reachable!"
         break
     else
-        echo "Waiting for container..."
+        echo "Waiting for MariaDB..."
         sleep 2
     fi
 done
 
-echo "MySQL is up!"
-# if [ -f wp-config.php ]
+sleep 3
+echo "MariaDB is up!"
 
-wp config create --dbhost=$MYSQL_HOST --dbname=$MYSQL_DATABASE --dbuser=$MYSQL_USER \
---dbpass=$MYSQL_PASSWORD --locale=en_US --path=/srv/www/wordpress --allow-root
+if [ -f $WP_PATH/wp-config.php ]; then
 
-wp core install --url=localhost --title=Inception --admin_user=admin --admin_password=admin \
- --admin_email=inception@42.fr --allow-root --path=/srv/www/wordpress
+    echo "wp-config.php exists!"
 
-sed -i '/^listen = /s/.*/listen = 0.0.0.0:9000/' /etc/php/8.2/fpm/pool.d/www.conf
+else
+
+    wp config create --dbhost=$MYSQL_HOST --dbname=$MYSQL_DATABASE --dbuser=$MYSQL_USER \
+    --dbpass=$MYSQL_PASSWORD --locale=en_US --path=$WP_PATH --allow-root
+
+    wp core install --url=localhost --title=Inception --admin_user=admin --admin_password=admin \
+    --admin_email=inception@42.fr --allow-root --path=$WP_PATH
+
+
+fi
+
+sed -i '/^listen = /s/.*/listen = wordpress:9000/' /etc/php/8.2/fpm/pool.d/www.conf
 
 exec "$@"
